fix table types income/outcome value;

+++ console/migrations/m170312_073325_create_utcome_table.php
@@ -17,7 +17,7 @@ class m170312_073325_create_utcome_table extends Migration
             'title' => $this->string(50),
             'createdAt' => $this->datetime()->notNull(),
             'updatedAt' => $this->datetime(),
-            'value' => $this->integer()->notNull(),
+            'value' => $this->double()->notNull(),
         ]);
     }
